Fix function `fields_at_savepoint` in Python API
A bug when calling the function `fs_fields_at_savepoint_names` was
causing the list of fields saved at a savepoint to return wrong names
and potentially to segfault.

The problem was in the allocation of space for the field names: there
was no space for the null-termination. One more character is now
allocated for each string, solving the issue.

Example output before the fix (the fields `x`, `Q` and `u` are saved at
the requested savepoint):

    In [7]: ser.fields_at_savepoint(ser.savepoints[0])
    Out[7]: ['x', 'x', 'x']

Same thing after the fix:

    In [3]: ser.fields_at_savepoint(ser.savepoints[0])
    Out[3]: ['Q', 'u', 'x']